requirements.txt
# Web
flask>=2.0.1
gunicorn>=20.1.0
flask-assets>=2.0
webassets>=2.0

# YouTube и видео
yt-dlp>=2023.10.13
pydub>=0.25.1
opencv-python-headless>=4.5.3  # headless версия для Docker
ffmpeg-python>=0.2.0

# ML/AI
openai-whisper>=1.0.0
torch>=2.0.0,<3.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0,<5.0.0
sentence-transformers==2.5.1
scikit-learn>=0.24.0

# Google API
google-auth-oauthlib>=0.5.3
google-auth-httplib2>=0.1.0
google-auth>=2.0.0
google-api-python-client>=2.0.0

# PDF и форматирование
markdown2>=2.4.0
pdfkit>=1.0.0,<2.0.0
Pillow>=9.0.0

# Очередь задач
celery>=5.3.6,<6.0.0
redis>=5.0.1,<6.0.0
amqp>=5.2.0
kombu>=5.3.5

# Утилиты
python-dotenv>=0.19.0
psutil==5.9.8
pyyaml>=6.0.0
requests>=2.26.0
numpy>=1.21.2

